Rename fixture .server.js to just .js
